Isaac ROS 4.4

+++ nvblox_test_data/asset_scripts/quickstart.sh
@@ -12,16 +12,20 @@
 
 set -e
 
-[[ -z "$ISAAC_ROS_WS" ]] && echo "ERROR: ISAAC_ROS_WS is not set." && exit 1
+if [ -z "$ISAAC_ROS_WS" ] && [ -n "$METADATA_PATH" ]; then
+  ISAAC_ROS_WS="$(readlink -f "$(dirname "${METADATA_PATH}")/../../..")"
+elif [ -z "$ISAAC_ROS_WS" ]; then
+  echo "ERROR: ISAAC_ROS_WS is not set." && exit 1
+fi
 
 # The dataset name
-EXAMPLE_BAG_NAME="quickstart"
-ARCHIVE_FILENAME="${EXAMPLE_BAG_NAME}.tar.gz"
+ASSET_NAME="quickstart"
+EULA_URL="https://catalog.ngc.nvidia.com/orgs/nvidia/teams/isaac/resources/isaac_ros_nvblox_assets"
+ASSET_DIR="${ISAAC_ROS_WS}/isaac_ros_assets"
+ASSET_INSTALL_PATHS="${ASSET_DIR}/isaac_ros_nvblox/${ASSET_NAME}/metadata.yaml ${ASSET_DIR}/isaac_ros_nvblox/${ASSET_NAME}/rosbag2_2024_04_04-15_44_33_0.db3"
+ARCHIVE_FILENAME="${ASSET_NAME}.tar.gz"
 
-# NOTE: This is the same path that the isaac_ros_nvblox quickstart is downloaded to.
-ROSBAG_ASSET_DIR="${ISAAC_ROS_WS}/isaac_ros_assets"
-ASSET_INSTALL_PATHS="${ROSBAG_ASSET_DIR}/isaac_ros_nvblox/${EXAMPLE_BAG_NAME}/metadata.yaml ${ROSBAG_ASSET_DIR}/isaac_ros_nvblox/${EXAMPLE_BAG_NAME}/rosbag2_2024_04_04-15_44_33_0.db3"
-DOWNLOAD_PATH="/tmp/${ARCHIVE_FILENAME}"
+DOWNLOAD_PATH="${ASSET_DIR}/${ARCHIVE_FILENAME}"
 
 [[ $1 == "--print-install-paths" ]] && echo -n "$ASSET_INSTALL_PATHS" && exit 0
 
@@ -35,10 +39,25 @@ NGC_FILENAME=$ARCHIVE_FILENAME
 REQ_URL="https://api.ngc.nvidia.com/v2/resources/org/$NGC_ORG/team/$NGC_TEAM/$NGC_RESOURCE/$NGC_VERSION/files?redirect=true&path=$NGC_FILENAME"
 
 set -x
-mkdir -p "${ROSBAG_ASSET_DIR}"
 
-echo "Downloading test bag: $EXAMPLE_BAG_NAME"
-echo "From: $REQ_URL"
-curl -L "${REQ_URL}" -o "${DOWNLOAD_PATH}"
-tar -xf "${DOWNLOAD_PATH}" -C "${ROSBAG_ASSET_DIR}"
+rm -rf "${ASSET_DIR}/isaac_ros_nvblox"
+mkdir -p "${ASSET_DIR}/isaac_ros_nvblox/${ASSET_NAME}"
+
+if [[ -n "${ISAAC_ROS_ASSETS_TEST}" ]]; then
+  if [[ -z "${ISAAC_ROS_NVBLOX_TEST_DATA}" || ! -f "${ISAAC_ROS_NVBLOX_TEST_DATA}" ]]; then
+    echo "ERROR: Cache path ${ISAAC_ROS_NVBLOX_TEST_DATA} does not exist."
+    exit 1
+  fi
+  exit 0
+fi
+
+if [[ -n "${ISAAC_ROS_NVBLOX_TEST_DATA}" && -f "${ISAAC_ROS_NVBLOX_TEST_DATA}" ]]; then
+  echo "Copying artifact from ${ISAAC_ROS_NVBLOX_TEST_DATA} to ${DOWNLOAD_PATH}."
+  cp "${ISAAC_ROS_NVBLOX_TEST_DATA}" "${DOWNLOAD_PATH}"
+else
+  echo "Downloading test bag: $ASSET_NAME from ${REQ_URL}"
+  curl -L "${REQ_URL}" -o "${DOWNLOAD_PATH}"
+fi
+
+tar -xvf "${DOWNLOAD_PATH}" -C "${ASSET_DIR}"
 rm -f "${DOWNLOAD_PATH}"
